+++ src/test/java/com/arael82/challenge/artists/TestConstants.java
@@ -2,7 +2,7 @@ package com.arael82.challenge.artists;
 
 public class TestConstants {
 
-    public static final int TEST_ARTIST_ID = 1;
+    public static final Long TEST_ARTIST_ID = 1L;
     public static final String TEST_ARTIST_NAME = "testName";
 
 }
